refactor: move jemalloc code to influxdb3 serve
