+++ influxdb3_process/Cargo.toml
@@ -13,26 +13,12 @@ build = "Core"
 [dependencies]
 # Core Crates
 iox_time.workspace = true
-metric.workspace = true
-tokio_metrics_bridge.workspace = true
 
 # Crates.io dependencies
-tokio.workspace = true
 uuid.workspace = true
 
 [build-dependencies]
 cargo_metadata.workspace = true
-
-# Optional Dependencies
-[target.'cfg(not(target_env = "msvc"))'.dependencies]
-tikv-jemalloc-ctl = { version = "0.5.4", optional = true }
-tikv-jemallocator = { version = "0.5", optional = true, features = ["unprefixed_malloc_on_supported_platforms"] }
-
-[features]
-default = ["jemalloc_replacing_malloc"]
-
-# Use jemalloc as the allocator.
-jemalloc_replacing_malloc = ["tikv-jemallocator", "tikv-jemalloc-ctl"]
 
 [lints]
 workspace = true
